CAMEL-16556 Multiple AtlasMap endpoints may overwrite each others' propertiesFiles
Moved propertiesFile option from endpoint to component
