+++ components/camel-atlasmap/src/main/java/org/apache/camel/component/atlasmap/AtlasMapEndpoint.java
@@ -22,13 +22,11 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Properties;
 
 import io.atlasmap.api.AtlasContext;
 import io.atlasmap.api.AtlasContextFactory;
 import io.atlasmap.api.AtlasException;
 import io.atlasmap.api.AtlasSession;
-import io.atlasmap.core.DefaultAtlasContextFactory;
 import io.atlasmap.v2.Audit;
 import io.atlasmap.v2.DataSource;
 import io.atlasmap.v2.DataSourceType;
@@ -40,8 +38,6 @@ import org.apache.camel.component.ResourceEndpoint;
 import org.apache.camel.spi.UriEndpoint;
 import org.apache.camel.spi.UriParam;
 import org.apache.camel.support.MessageHelper;
-import org.apache.camel.support.ResourceHelper;
-import org.apache.camel.util.IOHelper;
 import org.apache.camel.util.ObjectHelper;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -63,8 +59,6 @@ public class AtlasMapEndpoint extends ResourceEndpoint {
     private AtlasContextFactory atlasContextFactory;
     private AtlasContext atlasContext;
 
-    @UriParam(label = "advanced")
-    private String propertiesFile;
     @UriParam
     private String sourceMapName;
     @UriParam
@@ -111,19 +105,6 @@ public class AtlasMapEndpoint extends ResourceEndpoint {
 
     public void setAtlasContext(AtlasContext atlasContext) {
         this.atlasContext = atlasContext;
-    }
-
-    /**
-     * The URI of the properties file which is used for AtlasContextFactory initialization.
-     * 
-     * @param file property file path
-     */
-    public void setPropertiesFile(String file) {
-        propertiesFile = file;
-    }
-
-    public String getPropertiesFile() {
-        return propertiesFile;
     }
 
     /**
@@ -239,7 +220,7 @@ public class AtlasMapEndpoint extends ResourceEndpoint {
             }
             // remove the header to avoid it being propagated in the routing
             incomingMessage.removeHeader(AtlasMapConstants.ATLAS_MAPPING);
-            return getOrCreateAtlasContextFactory().createContext(JSON, is);
+            return atlasContextFactory.createContext(JSON, is);
         } else if (getAtlasContext() != null) {
             // no mapping specified in header, and found an existing context
             return getAtlasContext();
@@ -250,33 +231,9 @@ public class AtlasMapEndpoint extends ResourceEndpoint {
             log.debug("Atlas mapping content read from resourceUri: {} for endpoint {}",
                     path, getEndpointUri());
         }
-        atlasContext = getOrCreateAtlasContextFactory().createContext(
+        atlasContext = atlasContextFactory.createContext(
                 path.toLowerCase().endsWith("adm") ? ADM : JSON, getResourceAsInputStream());
         return atlasContext;
-    }
-
-    private synchronized AtlasContextFactory getOrCreateAtlasContextFactory() throws Exception {
-        if (atlasContextFactory != null) {
-            return atlasContextFactory;
-        }
-
-        atlasContextFactory = DefaultAtlasContextFactory.getInstance();
-        atlasContextFactory.addClassLoader(getCamelContext().getApplicationContextClassLoader());
-        // load the properties from property file which may overrides the default ones
-        if (ObjectHelper.isNotEmpty(getPropertiesFile())) {
-            Properties properties = new Properties();
-            InputStream reader = ResourceHelper.resolveMandatoryResourceAsInputStream(getCamelContext(),
-                    getPropertiesFile());
-            try {
-                properties.load(reader);
-                log.info("Loaded the Atlas properties file {}", getPropertiesFile());
-            } finally {
-                IOHelper.close(reader, getPropertiesFile(), log);
-            }
-            log.debug("Initializing AtlasContextFactory with properties {}", properties);
-            atlasContextFactory.setProperties(properties);
-        }
-        return atlasContextFactory;
     }
 
     private void populateSourceDocuments(Exchange exchange, AtlasSession session) {
